chore(remix): refactor testing config utils
